Moved tests to integration folder
