+++ test/integration/ping-test.js
@@ -2,8 +2,9 @@ var chai = require("chai"),
   chaiHttp = require("chai-http"),
   expect = chai.expect;
 
-var config = require("../config/config");
-var server = require("../app/server")(config);
+var config = require("../../config/config");
+var server = require("../../app/server")(config);
+var pkg = require("../../package.json");
 
 
 chai.use(chaiHttp);
@@ -12,8 +13,7 @@ describe("Ping", function () {
 
   beforeEach(function () {
     server.start();
-    var pkg = {name: "rise-cache-v2", version: "1.0.0"}
-    require("../app/controllers/ping")(server.app, pkg);
+    require("../../app/controllers/ping")(server.app, pkg);
   });
 
   it("should return name and version", function (done) {
@@ -23,8 +23,8 @@ describe("Ping", function () {
       expect(res).to.have.status(200);
       expect(res).to.be.json;
       expect(res.body).to.be.a('object');
-      expect(res.body.name).to.be.equal('rise-cache-v2');
-      expect(res.body.version).to.be.equal('1.0.0');
+      expect(res.body.name).to.be.equal(pkg.name);
+      expect(res.body.version).to.be.equal(pkg.version);
       done();
     });
   });
